+++ edfi.fif.api/src/graphql/entities/survey/surveysummaryquestions.entity.ts
@@ -1,10 +1,10 @@
 import { ViewEntity, ViewColumn, PrimaryColumn } from 'typeorm';
-import SurveySummaryAnsweres from './surveysummaryansweres.entity';
+import SurveySummaryAnsweres from './surveysummaryanswers.entity';
 import SurveySummaryEntity from './surveysummary.entity';
 
 @ViewEntity({ schema: 'fif', name: 'surveysummaryquestions', synchronize: false })
 export default class SurveySummaryQuestionsEntity {
-  
+
   @ViewColumn()
   surveykey: number;
